Apps/capture-hanzi-streams/pyproject.toml
[project]
name = "capture-hanzi-streams"
version = "0.1.0"
description = "Capture Simplified Chinese YouTube captions by time segment"
readme = "README.md"
authors = [
    { name = "Brian Abbott", email = "briancabbott@gmail.com" }
]
requires-python = ">=3.11"
dependencies = [
    "opencc-python-reimplemented>=0.1.7",
    "youtube-transcript-api>=1.2.4",
]

[project.scripts]
capture-hanzi-streams = "capture_hanzi_streams:main"

[build-system]
requires = ["uv_build>=0.11.7,<0.12.0"]
build-backend = "uv_build"
